chore: Move RunningStateLike to common
