+++ packages/common/src/index.ts
@@ -10,6 +10,7 @@ export * from './loggers'
 export * from './pinning'
 export * from './doc-cache'
 export * from './unreachable-case-error'
+export * from './running-state-like'
 
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
