+++ packages/core/src/local-pin-api.ts
@@ -1,8 +1,7 @@
-import { PinApi } from '@ceramicnetwork/common';
+import { PinApi, RunningStateLike } from '@ceramicnetwork/common';
 import DocID from '@ceramicnetwork/docid';
 import { DiagnosticsLogger } from "@ceramicnetwork/common";
 import { Repository } from './state-management/repository';
-import { RunningStateLike } from './state-management/running-state';
 
 /**
  * PinApi for Ceramic core.
