+++ packages/core/src/state-management/state-manager.ts
@@ -2,8 +2,8 @@ import { Dispatcher } from '../dispatcher';
 import { PinStore } from '../store/pin-store';
 import { ExecutionQueue } from './execution-queue';
 import { commitAtTime, ConflictResolution } from '../conflict-resolution';
-import { AnchorService, AnchorStatus, DocOpts, UnreachableCaseError } from '@ceramicnetwork/common';
-import { RunningState, RunningStateLike } from './running-state';
+import { AnchorService, AnchorStatus, DocOpts, UnreachableCaseError, RunningStateLike } from '@ceramicnetwork/common';
+import { RunningState } from './running-state';
 import CID from 'cids';
 import { timeoutWith } from 'rxjs/operators';
 import { of, Subscription } from 'rxjs';
